+++ avsc/src/decorators/avroResponse.decorator.ts
@@ -1,4 +1,4 @@
-import {RESTClient, AdditionalInfoPropertyDescriptor} from '@anglr/rest';
+import {RESTClient, AdditionalInfoPropertyDescriptor, ResponseType} from '@anglr/rest';
 
 import {AvroResponseType} from '../avsc';
 
@@ -21,6 +21,9 @@ export function AvroResponse(namespace: string, typeName: string)
             name: typeName,
             namespace: namespace
         };
+
+        //TODO finish
+        (descriptor as any).responseType = ResponseType.ArrayBuffer;
         
         return descriptor;
     };