+++ src/rest.ts
@@ -3,3 +3,4 @@ export {Cache} from './rest/cache';
 export {Body,BaseUrl,DefaultHeaders,DELETE,GET,HEAD,Header,Headers,JsonContentType,ParameterTransform,Path,POST,Produces,PUT,Query,QueryObject,ResponseTransform,RESTClient,FullHttpResponse,ReportProgress, DisableInterceptor, ProgressIndicatorGroup} from './rest/common';
 export {RestTransferStateService} from './transferState/restTransferState.service';
 export {AdditionalInfoPropertyDescriptor} from './rest/additionalInfoPropertyDescriptor';
+export {HTTP_HEADER_ACCEPT, HTTP_HEADER_CONTENT_TYPE} from './rest/constants';