+++ src/rest/constants.ts
@@ -1,0 +1,9 @@
+/**
+ * Http header name for "Content-Type" header
+ */
+export const HTTP_HEADER_CONTENT_TYPE = 'content-type';
+
+/**
+ * Http header name for "Accept" header
+ */
+export const HTTP_HEADER_ACCEPT = 'accept';